Integrate mod files (#1)
* Include the Noita mod files into this repository

* Use CRLF line endings for mod distribution

* Update mod content to v1.1-beta release zip contents

The files were not checked in to the parent repo

* Attempt to fix line endings...

* Set up workspace / autoformat

* Do all the work in one commit ... whoops

* Add no streamer page with migration note. Unify pug layout.

+++ controllers/authController.js
@@ -26,16 +26,23 @@ exports.isLoggedIn = (req, res, next) => {
     }
 }
 
-exports.logOut = async (req, res) => {
+exports.logOut = async (req, res, next) => {
+    if (!req.user) {
+        res.redirect('/')
+        return
+    }
+
     const deleted = await Tokens.findOneAndDelete({ id: req.user.id })
     if (!deleted) {
         const err = new Error('Failed to delete refresh token.')
         throw err
     }
 
-    req.logout()
-    req.session.destroy((err) => {
-        res.sendStatus(204)
+    req.logout(function (err) {
+        if (err) {
+            return next(err)
+        }
+        res.redirect('/')
     })
 }
